feat/ui: merge mint modal with mint component
